requirements.txt
Flask==0.10.1
Jinja2==2.8
MarkupSafe==0.23
Werkzeug==0.11.4
argparse==1.2.1
gevent==1.0.2
greenlet==0.4.9
itsdangerous==0.24
locustio==0.7.3
msgpack-python==0.4.7
pyzmq==15.2.0
requests==2.9.1
wheel==0.29.0
wsgiref==0.1.2
